android_servers: Fix the duration argument for CPU_BOOST
Clear incompatible pointer conversion errors in 64bit

+++ services/core/jni/com_android_server_power_PowerManagerService.cpp
@@ -159,7 +159,7 @@ static void nativeSendPowerHint(JNIEnv *env, jclass clazz, jint hintId, jint dat
 static void nativeCpuBoost(JNIEnv *env, jobject clazz, jint duration) {
     // Tell the Power HAL to boost the CPU
     if (gPowerModule && gPowerModule->powerHint) {
-        gPowerModule->powerHint(gPowerModule, POWER_HINT_CPU_BOOST, (void *) duration);
+        gPowerModule->powerHint(gPowerModule, POWER_HINT_CPU_BOOST, (void *)(static_cast<int64_t>(duration)));
     }
 }
